migrate test_dpkg_lowpkg to pytest
